Revert "DACL 20140908"
This reverts commit 414e626c6fbc5f8bba8226e7e8a8781b6d548fcb.

+++ Logica/Controler/ControlEmpresa.php
@@ -2,51 +2,55 @@
 require_once 'SystemControl.php';
 require_once '../Model/Empresa.php';
 
-abstract class ControlEmpresa extends SystemControl {
-    private $empresa;
-    public function __construct($session) {
-        parent::__construct($session);
-        $this -> empresa = null;
-    }
+abstract class ControlEmpresa extends SystemControl
+{
+	private $empresa;
+	public function __construct($session){
+		parent::__construct($session);
+		$this->empresa = null;
+	}
 
-    public function __destruct() {
+    public function __destruct(){
         parent::__destruct();
     }
 
-    public function setEmpresa($arrempresa) {
-        $this -> empresa = new Empresa($arrempresa);
-    }
+	public function setEmpresa($arrempresa){
+		$this->empresa = new Empresa($arrempresa);		
+	}
 
-    /**
-     *$empresa deve estar cargado previamente para
-     *poder ser creada, actualizada o eliminada.
-     */
+	/**
+	 *$empresa deve estar cargado previamente para
+	 *poder ser creada, actualizada o eliminada.
+	 */
+	 
+	//crea una empresa
+	final public function crearEmpresa(){
+		if($this->empresa == null){
+			throw new Exception('Empresa sin datos');
+		}
+		DataAccess::write($this->empresa);
+		echo TRUE;
+	}
+	
+	//modifica una empresa existente
+	final protected function modificarEmpresa(){
+		if($this->empresa != null){
+			throw new Exception('Empresa sin datos');
+		}
+		DataAccess::update($this->empresa);	
+	}  
 
-    //crea una empresa
-    final public function crearEmpresa() {
-
-        Tools::thisException('Usuario sin privilejios', !parent::$roles['Empresa_reg']);
-        Tools::thisException('Empresa sin datos', $this -> empresa == null);
-
-        DataAccess::write($this -> empresa);
-        echo TRUE;
-    }
-
-    //modifica una empresa existente
-    final protected function modificarEmpresa() {    
-        Tools::thisException('Empresa sin datos',$this -> empresa != null);
-        DataAccess::update($this -> empresa);
-    }
-
-    //Elimina una empresa existente
-    final protected function eliminar() {
-        Tools::thisException('Empresa sin datos',$this -> empresa != null);
-        DataAccess::delete($this -> empresa);
-    }
-
-    final protected function consultarEmpresas() {
-        return DataAccess::selectWhere($this -> empresa);
-    }
-
+	//Elimina una empresa existente
+	final protected function eliminar(){
+		if($this->empresa != null){
+			throw new Exception('Empresa sin datos');
+		}
+		DataAccess::delete($this->empresa);	
+	}
+	
+	final protected function consultarEmpresas()
+	{
+		return DataAccess::selectWhere($this->empresa);
+	}
 }
 ?>
